% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmRegress.R
\name{covHuber}
\alias{covHuber}
\title{calculate covariance matrix based on huber loss}
\usage{
covHuber(x)
}
\arguments{
\item{x}{predictors}
}
\description{
calculate covariance matrix based on huber loss
}
