% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmRegress.R
\name{farmRegress}
\alias{farmRegress}
\title{farm regression}
\usage{
farmRegress(x, y, robust = T, type = c("regression", "classification"), ...)
}
\arguments{
\item{x}{predictors}

\item{y}{response}
}
\description{
farm regression
}
