% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmRegress.R
\name{findLambdaClass}
\alias{findLambdaClass}
\title{find B matrix such that X=BxF+e, i.e. lambda.hat}
\usage{
findLambdaClass(covx, N, P, k.factors)
}
\arguments{
\item{covx}{covariance of x}

\item{N}{row number of x}

\item{P}{col number of x}

\item{k.factors}{number of factors needed}
}
\description{
find B matrix such that X=BxF+e, i.e. lambda.hat
}
