% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmRegress.R
\name{findXStar}
\alias{findXStar}
\title{find x star such that U = X - BF under linear regression}
\usage{
findXStar(x, PF)
}
\arguments{
\item{x}{predictors}

\item{PF}{transformation of F}
}
\description{
find x star such that U = X - BF under linear regression
}
