% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmRegress.R
\name{findXStarClass}
\alias{findXStarClass}
\title{find U such that U = X - BF}
\usage{
findXStarClass(x, lambda.hat, F.hat)
}
\arguments{
\item{x}{predictors}

\item{lambda.hat}{estimation of B}

\item{F.hat}{estimation of F}
}
\description{
find U such that U = X - BF
}
