% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmRegress.R
\name{findFactorClass}
\alias{findFactorClass}
\title{find F such that X = BxF + e, i.e F.hat}
\usage{
findFactorClass(x, eigen.value, lambda.hat)
}
\arguments{
\item{x}{predictors}

\item{eigen.value}{covraince matrix of x}

\item{lambda.hat}{the estimation of B}
}
\description{
find F such that X = BxF + e, i.e F.hat
}
