% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmRegress.R
\name{findLambdaClass}
\alias{findLambdaClass}
\title{find B matrix such that X=BxF+e, i.e. lambda.hat}
\usage{
findLambdaClass(eigen.value, eigen.vector, k.factors)
}
\arguments{
\item{eigen.value}{eigen values of covariance matrix}

\item{eigen.vector}{eigen vectors of covariance matrix}

\item{k.factors}{number of factors needed}
}
\description{
find B matrix such that X=BxF+e, i.e. lambda.hat
}
